Parsing: creating task for GoodURL (#2)
* Parsing: creating task for GoodURL

* parsing.admin.GoodURLAdmin: added filter by row enabled
* parsing.models.Sitemap: added blank=true for row task
* parsing.models.GoodURL: fix method for creating task
* parsing.signals: fix pre_save and pre_delete signals
* pytest: deleted param for reused db

* Github pytest.ini: add django variable

+++ parsing/signals.py
@@ -12,33 +12,39 @@ logger = logging.getLogger(__name__)
 def site_update_or_create(sender, instance, using, **kwargs):
     logger.debug('Signal models.Site pre_save: %s', instance)
     instance.task = instance.task_create_or_update()
+    instance.schedule=instance.task.interval
 
 
 @receiver(pre_delete, sender=models.Site)
 def site_delete(sender, instance, using, **kwargs):
     logger.debug('Signal models.Site pre_delete: %s', instance)
-    instance.task.delete()
+    if instance.task:
+        instance.task.delete()
 
 
 @receiver(pre_save, sender=models.Sitemap)
 def sitemap_update_or_create(sender, instance, using, **kwargs):
     logger.debug('Signal models.sitemap pre_save: %s', instance)
     instance.task = instance.task_create_or_update()
+    instance.schedule=instance.task.interval
 
 
 @receiver(pre_delete, sender=models.Sitemap)
 def sitemap_delete(sender, instance, using, **kwargs):
     logger.debug('Signal models.Sitemap pre_delete: %s', instance)
-    instance.task.delete()
+    if instance.task:
+        instance.task.delete()
 
 
 @receiver(pre_save, sender=models.GoodURL)
 def goodurl_update_or_create(sender, instance, using, **kwargs):
     logger.debug('Signal model.GoodURL pre_save: %s', instance)
     instance.task = instance.task_create_or_update()
+    instance.schedule=instance.task.interval
 
 
 @receiver(pre_delete, sender=models.GoodURL)
 def goodurl_delete(sender, instance, using, **kwargs):
     logger.debug('Signal model.GoodURL pre_delete: %s', instance)
-    instance.task.delete()
+    if instance.task:
+        instance.task.delete()
